+++ ios/Runner/AppDelegate.m
@@ -98,7 +98,14 @@ UIInterfaceOrientationMask _interfaceOrientationToMask(UIInterfaceOrientation va
 	
 	// Setup supported & preffered orientation
 	_preferredInterfaceOrientation = UIInterfaceOrientationPortrait;
-	_supportedInterfaceOrientations = [NSSet setWithObject:@(_preferredInterfaceOrientation)];
+    //support only portrait
+    //	_supportedInterfaceOrientations = [NSSet setWithObject:@(_preferredInterfaceOrientation)];
+    // Support All orientations
+    NSMutableSet *initialOrientations = [NSMutableSet set];
+    [initialOrientations addObject:@(UIInterfaceOrientationPortrait)];
+    [initialOrientations addObject:@(UIInterfaceOrientationLandscapeLeft)];
+    [initialOrientations addObject:@(UIInterfaceOrientationLandscapeRight)];
+    _supportedInterfaceOrientations = [initialOrientations copy];
 
 	// Setup root ViewController
 	UIViewController *rootViewController = self.window.rootViewController;
@@ -108,6 +115,10 @@ UIInterfaceOrientationMask _interfaceOrientationToMask(UIInterfaceOrientation va
 	_navigationViewController.navigationBarHidden = YES;
 	[_navigationViewController setNeedsUpdateOfSupportedInterfaceOrientationsIfPossible];
 	_navigationViewController.delegate = self;
+
+    //At this point prompts the system to re-evaluate the supported orientations based on the values RootNavigationController will provide
+    //Call this AFTER _navigationViewController is initialized and its essential properties are set.
+    [_navigationViewController setNeedsUpdateOfSupportedInterfaceOrientationsIfPossible];
 
 	_navigationViewController.navigationBar.translucent = NO;
 	_navigationViewController.navigationBar.barTintColor = [UIColor inaColorWithHex:@"13294b"];