+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- The next version number should be the version already set in pubspec.yaml -->
 ### Added
 - Show events in the assistant as horizontal scroll [#5197](https://github.com/rokwire/illinois-app/issues/5197).
+- Enable Landscape Orientation [#1457](https://github.com/rokwire/illinois-app/issues/1457).
 
 ## [7.1.13] - 2025-06-02
 ### Fixed